Progress on LIAR Dataset
I attempted to tackle the Liar Dataset by learning how to use the BERT model, and the goal was to achieve good accuracy. In the end, I only managed to run 26% test accuracy by adding weight decay, decreased learning rate, and more epochs to make the model learn better. It was a much harder challenge to tackle and will continue to improve the model.